ml_project/requirements.txt
pkutils
jupyterlab==3.0.13
pandas==1.1.3
hydra-core==1.0.6
pandas-profiling==2.11.0
pytest==6.2.3
numpy==1.20.2
flake8
python-dotenv>=0.5.1
scikit-learn==0.24.2
setuptools~=56.0.0
omegaconf~=2.0.6